+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ## [Unreleased]
 ### Added
 - Add `use` for `EnvelopeTemplateEndpoint`
+- Add `AccountSettings.bankIdProduct`
+- Add `EnvelopeTemplate.bankIdScopes`
 
 ## [1.1.0] - 2021-06-25
 ### Added